Renamed utilities/strings to utilities/str
