+++ str/set.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package strings
+package str
 
 // StringSet is set implementation on top of map
 type StringSet struct {
